requirement.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
comtypes==1.4.16
cryptography==46.0.5
distro==1.9.0
Django==6.0.3
djangorestframework==3.16.1
google==3.0.0
google-auth==2.49.1
google-genai==1.67.0
gunicorn==25.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
mysqlclient==2.2.8
numpy==2.4.3
packaging==26.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pypiwin32==223
python-dotenv==1.2.2
pyttsx3==2.99
pywin32==311
requests==2.32.5
scipy==1.17.1
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.5
soupsieve==2.8.3
sqlparse==0.5.5
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
websockets==16.0
whisper==1.1.10
